+++ lib/Benchmark/Finder.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpBench;
+namespace PhpBench\Benchmark;
 
 use Symfony\Component\Finder\Finder;
